fix Models

+++ DMS/Models/ShowingOrderContentWithDrawDAO.cs
@@ -9,7 +9,7 @@ namespace DMS.Models
         public long ShowingOrderWithDrawId { get; set; }
         public long ShowingItemId { get; set; }
         public long UnitOfMeasureId { get; set; }
-        public long SalePrice { get; set; }
+        public decimal SalePrice { get; set; }
         public long Quantity { get; set; }
         public decimal Amount { get; set; }
